pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "your-scripts-package"
version = "0.1.0"
description = "Reusable scripts for the fantasy app"
authors = [{ name="Your Name", email="you@example.com" }]
dependencies = [
    "pymongo",
    "bson",
    "scramble-models @ git+ssh://git@github.com/JWehder/scramble-models.git",
    "selenium",
    "requests",
    "gspread",
    "python-dotenv",
    "pydantic"
]

[tool.setuptools]
package-dir = {"" = "scripts"}

[tool.setuptools.packages.find]
where = ["scripts"]